Changing montlo to fuseext2 (fixes #4)
This is a great use of FUSE and it allows to create filesystem images as a regular user.

There are now FUSE filesystems for ext2, FAT, NTFS, etc so there is no need for mountlo.

This modified recipe uses fuseext2 and it fixes some issues with the shell script in the presence of bootable flags.